+++ src/CreateNewPostPage/DetailPosting.jsx
