+++ src/CreateNewPostPage/MainPosting.jsx
